requirements.txt
django >= 1.4
django-mptt >= 0.5.4